Fix issues with saving controls

+++ TombEngine/Specific/configuration.h
@@ -64,7 +64,7 @@ struct GameConfiguration {
 	int MusicVolume;
 	int SfxVolume;
 	int SoundDevice;
-	byte KeyboardLayout[TEN::Input::NUM_CONTROLS];
+	short KeyboardLayout[TEN::Input::NUM_CONTROLS];
 	bool ControlMethod;
 	int JoyDuck;
 	int JoyAction;
